Denote release v1.7.1
Note that this project does not include the tests themselves. Instead, it provides infrastructure for an administrator to provide a Bamboo Executable that provides the tests.

Download the latest release of the Smack SINT Server Extensions project, which should be a JAR-file. This is the file that will be added as a Bamboo Executable capability.

The version number of this release mimics that of a release of the Smack SINT Server Extensions project, in which functionality is exposed that's now also available in this Bamboo project (enabling specific tests or specifications).